requirements.txt
aistudio-sdk==0.3.8
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astroid==4.0.4
asttokens==3.0.1
attrs==26.1.0
bce-python-sdk==0.9.67
black==26.3.1
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer>=2.0,<4.0
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
coverage==7.13.5
crc32c==2.8
cryptography==46.0.5
debugpy==1.8.20
decorator==5.2.1
dill==0.4.1
executing==2.2.1
filelock==3.25.2
flake8==7.3.0
fsspec==2026.2.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
imagesize==2.0.0
iniconfig==2.3.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
isort==8.0.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
modelscope==1.35.1
mpmath==1.3.0
mypy_extensions==1.1.0
narwhals==2.18.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.3
opencv-contrib-python==4.10.0.84
packaging==26.0
paddleocr==3.4.0
paddlex==3.4.2
pandas==2.3.3
parso==0.8.6
pathspec==1.0.4
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
prettytable==3.17.0
prompt_toolkit==3.0.52
protobuf==6.33.6
psutil==7.2.2
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==23.0.1
pyclipper==1.4.0
pycodestyle==2.14.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyflakes==3.4.0
Pygments==2.19.2
pylint==4.0.5
pypdf==6.9.2
pypdfium2==5.6.0
pytesseract==0.3.13
pytest==9.0.2
pytest-cov==7.1.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
pytokens==0.4.1
pytz==2026.1.post1
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.37.0
regex==2026.2.28
requests>=2.31,<3.0
rich==14.3.3
rpds-py==0.30.0
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==81.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
stack-data==0.6.3
streamlit==1.55.0
structlog==25.5.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
tomlkit==0.14.0
torch==2.11.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.12.0
urllib3>=2.0,<3.0
watchdog==6.0.0
wcwidth==0.6.0
wheel==0.46.3